Restructured to suit react convention. Added comments for testing application with IntelliJ IDEA
